backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0
python-docx==1.1.0
PyPDF2==3.0.1
google-genai==0.3.0
langchain==0.1.0
# Use compatible versions for Python 3.13
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
python-dotenv==1.0.0
